+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 - Enable context menu validation and added auto-scan option
 - Dynamic icon updates when in web view
 - Create chrome and firefox dist zip artifact
+- Add individual trust anchors list
 
 ## v0.1.2
 
-- Added support for Firefox
-- Added option to scan all media on a page
+- Add support for Firefox
+- Add option to scan all media on a page
 - Various updates
 
 ## v0.1.1